Added 102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,16 +6,16 @@
  */
 int main(void)
 {
-	unsigned long long n1 = 1;
-	unsigned long long n2 = 2;
-	unsigned long long current_fib;
+	unsigned long n1 = 1;
+	unsigned long n2 = 2;
+	unsigned long current_fib;
 	int i;
 
-	printf("%llu, %llu, ", n1, n2);
+	printf("%lu, %lu, ", n1, n2);
 	for (i = 1; i <= 48; i++)
 	{
 		current_fib = n1 + n2;
-		printf("%llu, ", (current_fib));
+		printf("%lu, ", (current_fib));
 		n1 = n2;
 		n2 = current_fib;
 	}
